apps/api/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
dj-database-url
python-dotenv
django-cors-headers
